Reorganize some library files and allow for verification without Stainless library
